Time: 51 ms (56.17%) | Memory: 59.4 MB (52.31%) - LeetSync

+++ 2665-minimum-time-to-repair-cars/minimum-time-to-repair-cars.cpp
@@ -4,7 +4,7 @@ public:
     bool is_it_Possible(vector<int>& ranks,int cars,ll mid){
         ll no_of_cars = 0 ;
         for(int i = 0 ;i < ranks.size() ;i ++){
-            no_of_cars += (ll)sqrt(mid/ranks[i]);
+            no_of_cars += sqrt(mid/ranks[i]);
         }
         return no_of_cars>=cars ;
     }
